Re #148 formal changes. Formally separated test for slider class and sliders widget (currently empty)
Use set instead of list to collect editing delegates. Does not make any difference.
